Move exceptions into a package
