+++ src/main/java/easton/Easton.java
@@ -3,6 +3,11 @@ package easton;
 import java.io.IOException;
 import java.util.ArrayList;
 
+import easton.exception.DateTimeFormatException;
+import easton.exception.EmptyDescriptionException;
+import easton.exception.IllegalActionException;
+import easton.exception.InvalidFormatException;
+import easton.exception.InvalidIndexException;
 import easton.model.Deadline;
 import easton.model.Event;
 import easton.model.Task;
