+++ src/main/java/easton/exception/EmptyDescriptionException.java
@@ -1,4 +1,4 @@
-package easton;
+package easton.exception;
 
 /**
  * Exception for an empty description.
@@ -7,7 +7,7 @@ public class EmptyDescriptionException extends Exception {
     /**
      * Construct a new exception with a generated detail message.
      */
-    EmptyDescriptionException() {
+    public EmptyDescriptionException() {
         super("The description cannot be empty!");
     }
 }
