+++ src/main/java/easton/exception/InvalidFormatException.java
@@ -1,4 +1,4 @@
-package easton;
+package easton.exception;
 
 /**
  * Exception for description is in the wrong format.
@@ -7,7 +7,7 @@ public class InvalidFormatException extends Exception {
     /**
      * Construct a new exception with a generated detail message.
      */
-    InvalidFormatException() {
+    public InvalidFormatException() {
         super("This description is in the wrong format!");
     }
 }
