+++ src/main/java/easton/exception/InvalidIndexException.java
@@ -1,4 +1,4 @@
-package easton;
+package easton.exception;
 
 /**
  * Exception for an invalid index is given.
@@ -7,7 +7,7 @@ public class InvalidIndexException extends Exception {
     /**
      * Construct a new exception with a generated detail message.
      */
-    InvalidIndexException(String input) {
+    public InvalidIndexException(String input) {
         super(input + " is an invalid index!");
     }
 }
