+++ src/test/java/easton/EastonTest.java
@@ -1,5 +1,8 @@
 package easton;
 
+import easton.exception.DateTimeFormatException;
+import easton.exception.EmptyDescriptionException;
+import easton.exception.InvalidFormatException;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
